feat: add new PubSubWithTestOptions class
