+++ src/index.ts
@@ -1,1 +1,2 @@
 export { PubSub } from './mock-pubsub';
+export { PubSubWithTestOptions } from './mock-pubsub-with-test-options';
